+++ thongssh_gtk/constants.py
@@ -1,7 +1,7 @@
 import sys
 import os
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 
 APP_ID = "com.example.thongssh"